ya_praktikum_aqa_projects/diploma_project/Diplom_2/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==8.1.1
requests==2.31.0
